requirements.txt
numpy==1.24.2
openai==0.28.1
opencv_contrib_python==4.8.1.78
pydub==0.25.1
pygame==2.5.2
python-dotenv==1.0.0
Requests==2.31.0
SpeechRecognition==3.10.0
